13/customizable record file name and location (#14)
Record file name and path is now customizable using rp-config.yml

+++ src/pomodoro_core/pomodoro.rs
@@ -126,7 +126,7 @@ impl<'a> Pomodoro<'a> {
     }
 
     fn convert_minutes_to_seconds(minutes: f32) -> u64 {
-        (minutes * (60 as f32)) as u64
+        (minutes * (60_f32)) as u64
     }
 }
